[#441] Refactor function names
* fixes #441
