+++ gradoop-flink/src/main/java/org/gradoop/flink/model/impl/operators/matching/single/cypher/functions/ProjectEdge.java
@@ -19,7 +19,7 @@ package org.gradoop.flink.model.impl.operators.matching.single.cypher.functions;
 
 import com.google.common.collect.Lists;
 import org.apache.flink.api.common.functions.RichMapFunction;
-import org.gradoop.common.model.impl.pojo.Vertex;
+import org.gradoop.common.model.impl.pojo.Edge;
 import org.gradoop.flink.model.impl.operators.matching.single.cypher.embeddings.Embedding;
 import org.gradoop.flink.model.impl.operators.matching.single.cypher.embeddings.IdEntry;
 import org.gradoop.flink.model.impl.operators.matching.single.cypher.utils.Projector;
@@ -29,30 +29,30 @@ import java.util.List;
 import java.util.Map;
 
 /**
- * Projects a Vertex by a set of properties.
- * Vertex -> Embedding(GraphElementEmbedding(Vertex))
+ * Projects an Edge by a set of properties.
+ * Edge -> Embedding(IdEntry(SrcID), GraphElementEntry(Edge), IdEntry(TargetID))
  */
-public class ProjectVertexFunction extends RichMapFunction<Vertex, Embedding> {
+public class ProjectEdge extends RichMapFunction<Edge, Embedding> {
   /**
    * Names of the properties that will be kept in the projection
    */
   private final Map<Integer, List<String>> propertyKeyMapping;
 
   /**
-   * Creates a new vertex projection function
-   * @param propertyKeys List of propertyKeys that will be kept in the projection
+   * Creates a new edge projection function
+   * @param propertyKeys List of property names that will be kept in the projection
    */
-  public ProjectVertexFunction(List<String> propertyKeys) {
+  public ProjectEdge(List<String> propertyKeys) {
     this.propertyKeyMapping = new HashMap<>();
-    propertyKeyMapping.put(0, propertyKeys);
+    propertyKeyMapping.put(1, propertyKeys);
   }
 
   @Override
-  public Embedding map(Vertex vertex) {
-    if (propertyKeyMapping.get(0).isEmpty()) {
-      return new Embedding(Lists.newArrayList(new IdEntry(vertex.getId())));
+  public Embedding map(Edge edge) {
+    if (propertyKeyMapping.get(1).isEmpty()) {
+      return new Embedding(Lists.newArrayList(new IdEntry(edge.getId())));
     } else {
-      return Projector.project(Embedding.fromVertex(vertex), propertyKeyMapping);
+      return Projector.project(Embedding.fromEdge(edge), propertyKeyMapping);
     }
   }
 }
