+++ gradoop-flink/src/main/java/org/gradoop/flink/model/impl/operators/matching/single/cypher/operators/physical/FilterAndProjectEmbeddings.java
@@ -20,7 +20,7 @@ package org.gradoop.flink.model.impl.operators.matching.single.cypher.operators.
 import org.apache.flink.api.java.DataSet;
 import org.gradoop.flink.model.impl.operators.matching.common.query.predicates.CNF;
 import org.gradoop.flink.model.impl.operators.matching.single.cypher.embeddings.Embedding;
-import org.gradoop.flink.model.impl.operators.matching.single.cypher.functions.FilterAndProjectEmbeddingFunction;
+import org.gradoop.flink.model.impl.operators.matching.single.cypher.functions.FilterAndProjectEmbedding;
 
 import java.util.List;
 import java.util.Map;
@@ -68,7 +68,7 @@ public class FilterAndProjectEmbeddings implements PhysicalOperator {
   @Override
   public DataSet<Embedding> evaluate() {
     return input.flatMap(
-      new FilterAndProjectEmbeddingFunction(predicates, propertyKeyMapping, columnMapping)
+      new FilterAndProjectEmbedding(predicates, propertyKeyMapping, columnMapping)
     );
   }
 }
