+++ gradoop-flink/src/main/java/org/gradoop/flink/model/impl/operators/matching/single/cypher/operators/physical/FilterAndProjectVertices.java
@@ -21,7 +21,7 @@ import org.apache.flink.api.java.DataSet;
 import org.gradoop.common.model.impl.pojo.Vertex;
 import org.gradoop.flink.model.impl.operators.matching.common.query.predicates.CNF;
 import org.gradoop.flink.model.impl.operators.matching.single.cypher.embeddings.Embedding;
-import org.gradoop.flink.model.impl.operators.matching.single.cypher.functions.FilterAndProjectVertexFunction;
+import org.gradoop.flink.model.impl.operators.matching.single.cypher.functions.FilterAndProjectVertex;
 
 import java.util.List;
 
@@ -60,7 +60,7 @@ public class FilterAndProjectVertices implements PhysicalOperator {
   @Override
   public DataSet<Embedding> evaluate() {
     return input.flatMap(
-      new FilterAndProjectVertexFunction(predicates, propertyKeys)
+      new FilterAndProjectVertex(predicates, propertyKeys)
     );
   }
 }
